+++ snooty/parser.py
@@ -8,21 +8,20 @@ import errno
 import pwd
 import subprocess
 import threading
-import yaml
 from copy import deepcopy
 from dataclasses import dataclass
 from functools import partial
 from pathlib import Path, PurePath
 from typing import Any, Dict, MutableSequence, Tuple, Optional, Set, List, Iterable
+from docutils.nodes import make_id
 from typing_extensions import Protocol
 import docutils.utils
 import watchdog.events
 import networkx
 
-from .flutter import check_type, LoadError
 from . import n, gizaparser, rstparser, util
 from .gizaparser.nodes import GizaCategory
-from .gizaparser.published_branches import PublishedBranches
+from .gizaparser.published_branches import PublishedBranches, parse_published_branches
 from .postprocess import DevhubPostprocessor, Postprocessor
 from .util import RST_EXTENSIONS
 from .types import (
@@ -49,8 +48,7 @@ from .diagnostics import (
     CannotOpenFile,
     InvalidURL,
     InvalidTableStructure,
-    UnmarshallingError,
-    ErrorParsingYAMLFile,
+    MalformedGlossary,
 )
 
 # XXX: Work around to get snooty working with Python 3.8 until we can fix
@@ -63,7 +61,7 @@ logger = logging.getLogger(__name__)
 
 @dataclass
 class _DefinitionListTerm(n.InlineParent):
-    """A private node used for internal book-keeping that should not be exported to the AST."""
+    """A vate node used for internal book-keeping that should not be exported to the AST."""
 
     __slots__ = ()
     type = "definition_list_term"
@@ -331,6 +329,37 @@ class JSONVisitor:
                     repr(top_of_state),
                     repr(popped),
                 )
+
+        if (
+            isinstance(popped, n.Directive)
+            and f"{popped.domain}:{popped.name}" == ":glossary"
+        ):
+
+            definition_list = next(popped.get_child_of_type(n.DefinitionList), None)
+
+            if definition_list is None:
+                return
+
+            if len(popped.children) != 1:
+                self.diagnostics.append(MalformedGlossary(util.get_line(node)))
+                return
+
+            if popped.options.get("sorted", False):
+                definition_list.children = sorted(
+                    definition_list.children,
+                    key=lambda DefinitionListItem: "".join(
+                        term.get_text() for term in DefinitionListItem.term
+                    ),
+                )
+
+            for item in definition_list.get_child_of_type(n.DefinitionListItem):
+                term_text = "".join(term.get_text() for term in item.term)
+                term_identifier = make_id(term_text)
+                identifier = n.TargetIdentifier(item.start, [], [term_identifier])
+                identifier.children = item.term[:]
+                target = n.InlineTarget(item.start, [], "std", "term", None)
+                target.children = [identifier]
+                item.term.append(target)
 
     def handle_directive(
         self, node: docutils.nodes.Node, line: int
@@ -704,28 +733,13 @@ class _Project:
     def get_parsed_branches(
         self
     ) -> Tuple[Optional[PublishedBranches], List[Diagnostic]]:
-        path = self.config.root
         try:
-            with path.joinpath("published-branches.yaml").open(encoding="utf-8") as f:
-                data = yaml.safe_load(f)
-                try:
-                    result = check_type(PublishedBranches, data)
-                    return result, []
-                except LoadError as err:
-                    line: int = getattr(err.bad_data, "_start_line", 0) + 1
-                    error_node: Diagnostic = UnmarshallingError(str(err), line)
-                    return None, [error_node]
+            path = self.config.root
+            return parse_published_branches(
+                path.joinpath("published-branches.yaml"), self.config
+            )
         except FileNotFoundError:
             pass
-        except LoadError as err:
-            load_error_line: int = getattr(err.bad_data, "_start_line", 0) + 1
-            load_error_node: Diagnostic = UnmarshallingError(str(err), load_error_line)
-            return None, [load_error_node]
-        except yaml.error.MarkedYAMLError as err:
-            yaml_error_node: Diagnostic = ErrorParsingYAMLFile(
-                path, str(err), err.problem_mark.line
-            )
-            return None, [yaml_error_node]
         return None, []
 
     def get_fileid(self, path: PurePath) -> FileId: